removed movement updat from AbstractGameObject class

+++ core/src/com/metalmethodd/orbhive/Constants.java
@@ -17,9 +17,9 @@ public class Constants {
     public static final int SCREEN_PADDING_TOP = 0;
     public static final int SCREEN_PADDING_BOTTOM = 36;
 
-    public static final float BULLET_HEIGHT = 4;
-    public static final float BULLET_WIDTH = 7;
-    public static final float BULLET_SPEED = 9 ;
+    public static final int BULLET_HEIGHT = 4;
+    public static final int BULLET_WIDTH = 7;
+    public static final int BULLET_SPEED = 9 ;
 
     public static final int SIMPLE_ENEMY_WIDTH = 16;
     public static final int SIMPLE_ENEMY_HEIGHT = 16;
